handle multiple versions of python

+++ mirror-sdists.sh
@@ -5,23 +5,22 @@ set -o pipefail
 export PS4='+ ${BASH_SOURCE#$HOME/}:$LINENO \011'
 
 WORKDIR=$(realpath $(pwd)/work-dir)
-if [ -d $WORKDIR ]; then
-    echo "Clean up $WORKDIR first"
-    exit 1
-fi
 mkdir -p $WORKDIR
 
+PYTHON=${PYTHON:-python3.9}
+PYTHON_VERSION=$($PYTHON --version | cut -f2 -d' ')
+
 # Redirect stdout/stderr to logfile
-logfile="$WORKDIR/mirror-sdists.log"
+logfile="$WORKDIR/mirror-sdists-${PYTHON_VERSION}.log"
 exec > >(tee "$logfile") 2>&1
 
 TOPLEVEL="${1:-langchain}"
 
-VENV=$WORKDIR/venv
-PYTHON=python3.9
+VENV=$WORKDIR/venv-${PYTHON_VERSION}
 
 WHEELS_REPO=$(realpath $(pwd)/wheels-repo)
 SDISTS_REPO=$(realpath $(pwd)/sdists-repo)
+BUILD_ORDER=${SDISTS_REPO}/build-order-${PYTHON_VERSION}.json
 
 setup() {
   if [ ! -d $VENV ]; then
@@ -43,11 +42,17 @@ setup() {
       packaging
 }
 
+already_processed() {
+  local req="$1"
+
+  jq .[].req < "${BUILD_ORDER}" | grep "$req"
+}
+
 add_to_build_order() {
   local type="$1"; shift
   local req="$1"; shift
   local why="$1"; shift
-  jq --argjson obj "{\"type\":\"${type}\",\"req\":\"${req//\"/\'}\",\"why\":\"${why}\"}" '. += [$obj]' ${SDISTS_REPO}/build-order.json > tmp.$$.json && mv tmp.$$.json ${SDISTS_REPO}/build-order.json
+  jq --argjson obj "{\"type\":\"${type}\",\"req\":\"${req//\"/\'}\",\"why\":\"${why}\"}" '. += [$obj]' ${BUILD_ORDER} > tmp.$$.json && mv tmp.$$.json ${BUILD_ORDER}
 }
 
 update_mirror() {
@@ -69,7 +74,7 @@ download_sdist() {
 
 get_downloaded_sdist() {
     local input=$1
-    grep Saved $input | cut -d ' ' -f 2-
+    grep -E '(Existing|Saved)' $input | cut -d ' ' -f 2-
 }
 
 safe_install() {
@@ -101,9 +106,19 @@ collect_build_requires() {
   local sdist="$1"; shift
   local why="$1"; shift
 
+  # Check if we have already processed this requirement to avoid cycles.
+  if already_processed "$req"; then
+      echo "$req has already been seen"
+      return
+  fi
+
   local next_why=$(basename ${sdist} .tar.gz)
 
   local unpack_dir=${WORKDIR}/$(basename ${sdist} .tar.gz)
+  # If the sdist was already unpacked we may have been doing the
+  # analysis with a different Python version, so remove the directory
+  # to ensure we get clean results this time.
+  rm -rf "${unpack_dir}"
   mkdir -p "${unpack_dir}"
   tar -C $unpack_dir -xvzf $sdist
   # We can't always predict what case will be used in the directory
@@ -122,6 +137,15 @@ collect_build_requires() {
   (cd ${extract_dir} && $PYTHON $extract_script --build-system "${req}") | tee "${build_system_deps}"
 
   cat "${build_system_deps}" | while read -r req_iter; do
+
+      # Check if we have already processed this requirement to avoid
+      # cycles. Do this before recursing to avoid adding an item to the
+      # build order list multiple times.
+      if already_processed "$req_iter"; then
+          echo "$req_iter has already been seen"
+          continue
+      fi
+
       download_output=${WORKDIR}/download-$(${parse_script} "${req_iter}").log
       download_sdist "${req_iter}" | tee $download_output
       local req_sdist=$(get_downloaded_sdist $download_output)
@@ -141,6 +165,15 @@ collect_build_requires() {
   (cd ${extract_dir} && $PYTHON $extract_script --build-backend "${req}") | tee "${build_backend_deps}"
 
   cat "${build_backend_deps}" | while read -r req_iter; do
+
+    # Check if we have already processed this requirement to avoid
+    # cycles. Do this before recursing to avoid adding an item to the
+    # build order list multiple times.
+    if already_processed "$req_iter"; then
+        echo "$req_iter has already been seen"
+        continue
+    fi
+
     download_output=${WORKDIR}/download-$(${parse_script} "${req_iter}").log
     download_sdist "${req_iter}" | tee $download_output
     local req_sdist=$(get_downloaded_sdist $download_output)
@@ -164,6 +197,15 @@ collect_build_requires() {
   (cd ${extract_dir} && $PYTHON $extract_script "${req}") | tee "${normal_deps}"
 
   cat "${normal_deps}" | while read -r req_iter; do
+
+    # Check if we have already processed this requirement to avoid
+    # cycles. Do this before recursing to avoid adding an item to the
+    # build order list multiple times.
+    if already_processed "$req_iter"; then
+        echo "$req_iter has already been seen"
+        continue
+    fi
+
     download_output=${WORKDIR}/download-$(${parse_script} "${req_iter}").log
     download_sdist "${req_iter}" | tee $download_output
     local req_sdist=$(get_downloaded_sdist $download_output)
@@ -192,10 +234,9 @@ trap on_exit EXIT SIGINT SIGTERM
 
 setup
 
-rm -rf ${SDISTS_REPO}/; mkdir -p ${SDISTS_REPO}/downloads/
-echo -n "[]" > ${SDISTS_REPO}/build-order.json
+mkdir -p ${SDISTS_REPO}/downloads/
+echo -n "[]" > ${BUILD_ORDER}
 
-rm -rf "${WHEELS_REPO}"
 mkdir -p "${WHEELS_REPO}/downloads"
 start_wheel_server
